+++ lib/dialogs/checkbox_picker_dialog.dart
@@ -1,21 +1,21 @@
 // Copyright (c) 2018, codegrue. All rights reserved. Use of this source code
 // is governed by the MIT license that can be found in the LICENSE file.
 
-import 'package:flutter_material_pickers/pickers/checkbox_picker.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_material_pickers/pickers/checkbox_picker.dart';
 
-import 'responsive_dialog.dart';
+import '../flutter_material_pickers.dart';
 import '../interfaces/common_dialog_properties.dart';
+import 'responsive_dialog.dart';
 
 /// This is a support widget that returns an Dialog with checkboxes as a Widget.
 /// It is designed to be used in the showDialog method of other fields.
-class CheckboxPickerDialog extends StatefulWidget
-    implements ICommonDialogProperties {
+class CheckboxPickerDialog<T> extends StatefulWidget implements ICommonDialogProperties {
   CheckboxPickerDialog({
     this.title,
     required this.items,
-    required this.values,
-    required this.initialValues,
+    required this.selectedItems,
+    this.transformer,
     this.headerColor,
     this.headerTextColor,
     this.backgroundColor,
@@ -27,9 +27,9 @@ class CheckboxPickerDialog extends StatefulWidget
   });
 
   // Variables
-  final List<String> items;
-  final List<String> values;
-  final List<String>? initialValues;
+  final List<T> items;
+  final List<T>? selectedItems;
+  final Transformer<T>? transformer;
   @override
   final String? title;
   @override
@@ -50,19 +50,16 @@ class CheckboxPickerDialog extends StatefulWidget
   final String? cancelText;
 
   @override
-  State<CheckboxPickerDialog> createState() =>
-      _CheckboxPickerDialogState(initialValues);
+  State<CheckboxPickerDialog> createState() => _CheckboxPickerDialogState<T>(selectedItems);
 }
 
-class _CheckboxPickerDialogState extends State<CheckboxPickerDialog> {
-  _CheckboxPickerDialogState(List<String>? initialValues) {
+class _CheckboxPickerDialogState<T> extends State<CheckboxPickerDialog<T>> {
+  _CheckboxPickerDialogState(List<T>? selectedItems) {
     // make a shallow copy so we don't modify the original list
-    selectedValues = (initialValues == null)
-        ? List<String>.empty(growable: true)
-        : List<String>.from(initialValues);
+    this.selectedItems = (selectedItems == null) ? List<T>.empty(growable: true) : List<T>.from(selectedItems);
   }
 
-  late List<String> selectedValues;
+  late List<T> selectedItems;
 
   @override
   Widget build(BuildContext context) {
@@ -77,12 +74,12 @@ class _CheckboxPickerDialogState extends State<CheckboxPickerDialog> {
       maxShortSide: widget.maxLongSide,
       confirmText: widget.confirmText,
       cancelText: widget.cancelText,
-      child: CheckboxPicker(
+      child: CheckboxPicker<T>(
         items: widget.items,
-        values: widget.values,
-        selectedValues: selectedValues,
+        selectedItems: selectedItems,
+        transformer: widget.transformer,
       ),
-      okPressed: () => Navigator.of(context).pop(selectedValues),
+      okPressed: () => Navigator.of(context).pop(selectedItems),
     );
   }
 }