+++ lib/dialogs/scroll_picker_dialog.dart
@@ -4,18 +4,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_material_pickers/pickers/scroll_picker.dart';
 
-import 'responsive_dialog.dart';
+import '../flutter_material_pickers.dart';
 import '../interfaces/common_dialog_properties.dart';
+import 'responsive_dialog.dart';
 
 /// This is a support widget that returns an Dialog with checkboxes as a Widget.
 /// It is designed to be used in the showDialog method of other fields.
-class ScrollPickerDialog extends StatefulWidget
-    implements ICommonDialogProperties {
+class ScrollPickerDialog<T> extends StatefulWidget implements ICommonDialogProperties {
   ScrollPickerDialog({
     this.title,
     required this.items,
-    required this.values,
-    this.initialValue,
+    required this.selectedItem,
+    this.transformer,
     this.headerColor,
     this.headerTextColor,
     this.backgroundColor,
@@ -28,9 +28,9 @@ class ScrollPickerDialog extends StatefulWidget
   });
 
   // Variables
-  final List<String> items;
-  final List<String> values;
-  final String? initialValue;
+  final List<T> items;
+  final T? selectedItem;
+  final Transformer<T>? transformer;
   @override
   final String? title;
   @override
@@ -53,14 +53,13 @@ class ScrollPickerDialog extends StatefulWidget
   final bool showDivider;
 
   @override
-  State<ScrollPickerDialog> createState() =>
-      _ScrollPickerDialogState(initialValue);
+  State<ScrollPickerDialog> createState() => _ScrollPickerDialogState<T>(selectedItem ?? items[0]);
 }
 
-class _ScrollPickerDialogState extends State<ScrollPickerDialog> {
-  _ScrollPickerDialogState(this.selectedValue);
+class _ScrollPickerDialogState<T> extends State<ScrollPickerDialog<T>> {
+  _ScrollPickerDialogState(this.selectedItem);
 
-  String? selectedValue;
+  T selectedItem;
 
   @override
   Widget build(BuildContext context) {
@@ -75,14 +74,14 @@ class _ScrollPickerDialogState extends State<ScrollPickerDialog> {
       maxShortSide: widget.maxLongSide,
       confirmText: widget.confirmText,
       cancelText: widget.cancelText,
-      child: ScrollPicker(
+      child: ScrollPicker<T>(
         items: widget.items,
-        values: widget.values,
-        initialValue: selectedValue,
+        selectedItem: selectedItem,
         showDivider: widget.showDivider,
-        onChanged: (value) => setState(() => selectedValue = value),
+        onChanged: (value) => setState(() => selectedItem = value),
+        transformer: widget.transformer,
       ),
-      okPressed: () => Navigator.of(context).pop(selectedValue),
+      okPressed: () => Navigator.of(context).pop(selectedItem),
     );
   }
 }